Separate file for each collection, tests reorganization, remove collection_std feature
