+++ module/core/collection_tools/Cargo.toml
@@ -42,7 +42,6 @@ default = [
   "prelude",
   "collection_constructors",
   "collection_into_constructors",
-  "collection_std",
 ]
 
 full = [
@@ -50,7 +49,6 @@ full = [
   "prelude",
   "collection_constructors",
   "collection_into_constructors",
-  "collection_std",
 ]
 
 enabled = []
@@ -60,9 +58,7 @@ prelude = []
 collection_constructors = []
 # Collection constructors, using `into()` under the hood, like `into_hmap!( "key" => "val" )`
 collection_into_constructors = []
-# STD collection for no_std.
-collection_std = []
-# qqq : is this feature used? seems not. if yes, what is it responsible for? discuss
+# qqq : is this feature used? seems not. if yes, what is it responsible for? discuss -- not needed, removed
 
 [dependencies]
