+++ module/core/collection_tools/src/lib.rs
@@ -4,34 +4,13 @@
 #![ doc( html_root_url = "https://docs.rs/collection_tools/latest/collection_tools/" ) ]
 #![ doc = include_str!( concat!( env!( "CARGO_MANIFEST_DIR" ), "/", "Readme.md" ) ) ]
 
-// qqq : make subdirectory for each container
+// qqq : make subdirectory for each container -- done
 
-// qqq : move out of lib.rs file
-/// Not meant to be called directly.
-#[ doc( hidden ) ]
-#[ macro_export( local_inner_macros ) ]
-macro_rules! count
-{
-  ( @single $( $x : tt )* ) => ( () );
+// qqq : move out of lib.rs file -- moved to `collections.rs`
 
-  (
-    @count $( $rest : expr ),*
-  )
-  =>
-  (
-    < [ () ] >::len( &[ $( count!( @single $rest ) ),* ] )
-  );
-}
-
-/// Macros to construct the collections.
-/// Basically a tweaked version of `literally` crate but using `alloc` / `hashbrown` instead of `std`
-#[ cfg( all( feature = "enabled", feature = "collection_constructors" ) ) ]
-pub mod constructors;
-
-/// Macros to construct the collections, using `.into()` under the hood.
-/// Often requires explicitly specifying type to cast to.
-#[ cfg( all( feature = "enabled", feature = "collection_into_constructors" ) ) ]
-pub mod into_constructors;
+/// Module containing all collection macros
+#[ cfg( feature = "enabled" ) ]
+pub mod collections;
 
 /// Namespace with dependencies.
 #[ cfg( feature = "enabled" ) ]
@@ -57,15 +36,12 @@ pub mod protected
   #[ allow( unused_imports ) ]
   pub use super::orphan::*;
 
-  // #[ cfg( feature = "use_alloc" ) ]
   extern crate alloc;
 
-  // #[ cfg( feature = "use_alloc" ) ]
   #[ doc( inline ) ]
   #[ allow( unused_imports ) ]
   pub use alloc::vec::Vec;
 
-  // #[ cfg( feature = "use_alloc" ) ]
   #[ doc( inline ) ]
   #[ allow( unused_imports ) ]
   pub use alloc::collections::{ BinaryHeap, BTreeMap, BTreeSet, LinkedList, VecDeque };
@@ -74,7 +50,7 @@ pub mod protected
   #[ cfg( feature = "use_alloc" ) ]
   #[ doc( inline ) ]
   #[ allow( unused_imports ) ]
-  pub use hashbrown::{ HashMap, HashSet };
+  pub use super::dependency::hashbrown::{ HashMap, HashSet };
 
   #[ cfg( not( feature = "no_std" ) ) ]
   #[ doc( inline ) ]
@@ -106,40 +82,30 @@ pub mod exposed
 pub mod prelude
 {
 
-  #[ cfg( feature = "collection_constructors" ) ]
-  #[ doc( inline ) ]
-  #[ allow( unused_imports ) ]
-  pub use super::constructors::*;
+  // qqq : for Anton : uncomment, make it working and cover by tests
+  // #[ cfg( feature = "prelude" ) ]
+  // #[ doc( inline ) ]
+  // #[ allow( unused_imports ) ]
+  // pub use crate::
+  // {
+  //   HashMap as Map,
+  //   HashSet as Set,
+  //   HashMap,
+  //   HashSet,
+  //   VecDeque,
+  //   BTreeMap,
+  //   BTreeSet,
+  //   BinaryHeap,
+  //   LinkedList,
+  // };
 
-  #[ cfg( feature = "collection_into_constructors" ) ]
-  #[ doc( inline ) ]
-  #[ allow( unused_imports ) ]
-  pub use super::into_constructors::*;
-
-// qqq : for Antont : uncomment, make it working and cover by tests
-//   #[ cfg( feature = "prelude" ) ]
-//   #[ doc( inline ) ]
-//   #[ allow( unused_imports ) ]
-//   pub use std::collections::
-//   {
-//     HashMap as Map,
-//     HashSet as Set,
-//     HashMap,
-//     HashSet,
-//     VecDeque,
-//     BTreeMap,
-//     BTreeSet,
-//     BinaryHeap,
-//     LinkedList,
-//   };
-//
-//   #[ cfg( feature = "prelude" ) ]
-//   #[ doc( inline ) ]
-//   #[ allow( unused_imports ) ]
-//   pub use std::vec::
-//   {
-//     Vec,
-//     Vec as DynArray,
-//   };
+  // #[ cfg( feature = "prelude" ) ]
+  // #[ doc( inline ) ]
+  // #[ allow( unused_imports ) ]
+  // pub use crate::
+  // {
+  //   Vec,
+  //   Vec as DynArray,
+  // };
 
 }
