--- module/core/collection_tools/src/vec.rs
@@ -1,2 +1,0 @@
-
-pub use core::slice::Iter
