+++ module/core/collection_tools/tests/inc/mod.rs
@@ -1,16 +1,15 @@
-#[ allow( unused_imports ) ]
 use super::*;
 
-#[ cfg( any( feature = "collection_into_constructors") ) ]
-mod into_constructors;
-
-#[ cfg( any( feature = "collection_constructors" ) ) ]
-mod constructors;
-
-#[ cfg( any( feature = "collection_std" ) ) ]
-mod reexport;
+mod bmap;
+mod bset;
+mod heap;
+mod hmap;
+mod hset;
+mod list;
+mod vec;
+mod vecd;
 
 mod components;
 
-// qqq : make subdirectory for each container
-// qqq : don't put tests otsude of directory `inc`
+// qqq : make subdirectory for each container -- done
+// qqq : don't put tests otsude of directory `inc` -- done
